requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
DateTime==5.5
idna==3.10
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
urllib3==2.3.0
zope.interface==7.2
